requirements.txt
fastapi
uvicorn
requests
python-dotenv
apscheduler
gspread
oauth2client
supabase
openai
streamlit
pandas
plotly
